requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.16.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.10.0
uvicorn==0.29.0
websockets==12.0
